chore(#9): Update tests to 100% coverage
There were two untested parts:
1. register_task/3 when the last argument couldn't be parsed.
2. ?LOG_ERROR macro call.

The first one was straightforward.

For the second one the problem was that the macro expands to a case and
the tests were hitting only one of the two clauses. The workarounds I
found for this were:
1. The implemented one, do a test with the logging disabled.
2. Change the macro usage to logger:error/2.
3. Ensure that the ?LOG_ERROR macro only expands to one line.

I was not specially happy with any of these workarounds so I settled on
the only one that didn't change the source code.

Closes #9.